requirements.txt
vosk
sounddevice
pyttsx3
google-generativeai
requests
openai
shutil